backend/requirements.txt
alembic==1.8.1
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==3.24.1
flask-marshmallow==0.10.1
Flask-Migrate==4.0.0
Flask-SQLAlchemy==2.5.1
idna==3.3
importlib-metadata==5.1.0
importlib-resources==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.2.4
MarkupSafe==2.0.1
marshmallow==3.3.0
marshmallow-sqlalchemy==0.21.0
PyJWT==1.7.1
PyMySQL==1.0.2
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
urllib3==1.26.9
Werkzeug==2.0.2
zipp==3.11.0